requirements.txt
asgiref==3.6.0
autopep8==2.0.1
av==10.0.0
bitstring==4.0.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colour==0.1.5
cryptography==39.0.2
cv==1.0.0
Django==4.1.7
ecdsa==0.18.0
esptool==4.5.1
ffmpeg==1.4
ink==0.0.1
mutagen==1.46.0
noise==1.2.2
numpy==1.25.1
opencv-python==4.7.0.72
Pillow==10.0.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodomex==3.17
PyMySQL==1.0.2
PyNaCl==1.5.0
pypng==0.20220715.0
pyserial==3.5
qrcode==7.4.2
reedsolo==1.6.0
shiv==1.0.3
six==1.16.0
sqlparse==0.4.3
tk==0.1.0
tkmacosx==1.0.5
tkvideoplayer==2.3
typing_extensions==4.5.0
websockets==11.0.1
youtube-dl==2021.12.17
yt-dlp==2023.3.4
